Enable a C preprocessing program to be used on Fortran code.
Some Fortran compilers cannot do C preprocessing (or have an
exceptionally broken implementation) so allow an external
C preprocessing tool to be be used and then run the Fortran compiler on
the preprocessed output.  This is only done if CPP is defined, otherwise
use the Fortran compiler to do the preprocessing and the compilation in
a single step.