Allow a different C/C++ compiler to be used to generate dependencies.
Not all C/C++ compilers have -MM -MT implemented.  In such cases allow
a different compiler (e.g. gcc or clang) to be used to generate the
dependencies.

This is useful on, for example, Blue Gene machines where the IBM xl
compilers must be used to compile the code but cannot generate the
dependencies in our required format.