Bugfix to md5 checking.
It turns out that md5sum from GNU coreutils 8.21 (on laptop) tolerates
one or two spaces between the md5 hash and the filename whereas
coreutils 8.12 (on work computer) does not.  Let's work around this by
being much more careful with how we create the md5sum string that needs
to be validated.